Split EtudeManager into 2 classes
 - one standard service
 - one twig extension
